Extracted selectors into its own thing
